(3 of 4) added pre-trained knowledge bases
The files are to be added up to doctaepupae.tar.gz - this file contains a swarm that only needs to be unpacked and linked with "./relink.sh".

Each insect has been individually trained. All in all, training was nearly two years on an AMD FX-8350 with 32 GB RAM.

Instead of using the insects as a "swarm", each insect can also be used just as a "personality" delivering a non-empty "swarmdata.txt" file.